refactor: move the vacation and group permission guards into the services layer
The vacation transition module is going to live in the services layer and
needs these guards, so they move down ahead of it — the services layer
imports nothing from controllers today and must not start now.

Import paths only: no function body, signature or exported name changes.
The vacation permission resolver moves whole, list variant included.

Refs #106
