+++ src/controllers/group/tests/handlePutGroupHolidayCountry.test.ts
@@ -14,7 +14,7 @@ const {
 
 vi.mock("../../../middleware/authSession.js", () => ({ getAuth: vi.fn() }));
 
-vi.mock("../../groupUser/utils.js", () => ({
+vi.mock("../../../services/groupUser/groupAccess.js", () => ({
   assertGroupAdmin: mockAssertGroupAdmin,
 }));
