+++ src/controllers/vacation/handlePostVacationComment.ts
@@ -7,7 +7,7 @@ import { db } from "../../db/db.js";
 import { generateRandomUUID } from "../../utils/generateUUID.js";
 import { vacationEventType } from "../../db/schema/vacation-event-schema.js";
 import { notifyVacationComment } from "../../services/vacation/vacationNotifier.js";
-import { resolveVacationPermissions } from "./utils.js";
+import { resolveVacationPermissions } from "../../services/vacation/vacationPermissions.js";
 import { assertGroupWritable } from "../../services/billing/guards.js";
 import type { ValidatedCommentVacationType } from "../../services/vacation/types.js";
