+++ src/controllers/vacation/tests/handleBulkCancelVacation.test.ts
@@ -26,7 +26,7 @@ const {
   mockResolveGroupAdmin: vi.fn(),
 }));
 
-vi.mock("../../groupUser/utils.js", () => ({
+vi.mock("../../../services/groupUser/groupAccess.js", () => ({
   resolveGroupAdmin: mockResolveGroupAdmin,
 }));
