+++ src/controllers/vacation/tests/handleDeleteVacation.test.ts
@@ -24,7 +24,7 @@ const {
   mockResolveGroupAdmin: vi.fn(),
 }));
 
-vi.mock("../../groupUser/utils.js", () => ({
+vi.mock("../../../services/groupUser/groupAccess.js", () => ({
   resolveGroupAdmin: mockResolveGroupAdmin,
 }));
