+++ src/services/vacation/decisionGuards.ts
@@ -1,7 +1,7 @@
 import AppError from "../../utils/appError.js";
 import type { DbTransaction } from "../../db/db.js";
-import { createDBServices } from "../../services/DBServices.js";
-import type { VacationType } from "../../services/vacation/types.js";
+import { createDBServices } from "../DBServices.js";
+import type { VacationType } from "./types.js";
 
 const services = createDBServices();
